requirements.txt
annotated-types==0.7.0
av==14.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gTTS==2.5.4
httplib2==0.22.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.3
onnxruntime==1.21.0
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
waitress==3.0.0
